+++ mod/src/main/java/basemod/interfaces/MaxHPChangeSubscriber.java
@@ -2,5 +2,14 @@ package basemod.interfaces;
 
 public interface MaxHPChangeSubscriber extends ISubscriber
 {
-	int receiveMapHPChange(int amount);
+	@Deprecated
+	default int receiveMapHPChange(int amount)
+	{
+		return receiveMaxHPChange(amount);
+	}
+
+	default int receiveMaxHPChange(int amount)
+	{
+		return amount;
+	}
 }